complete

+++ ImageCollectionApp/Services/StorageManager.swift
@@ -53,20 +53,20 @@ class StorageManager {
         }
     }
     
-    func save(_ newImage: Image) {
-//        guard let entityDescription = NSEntityDescription.entity(forEntityName: "SaveImage", in: viewContext) else { return }
-//        guard let image = NSManagedObject(entity: entityDescription, insertInto: viewContext) as? SaveImage else { return }
+    func save(_ newImage: Image?) {
         let image = SaveImage(context: viewContext)
-        image.image = newImage.urls.small
-        image.authorName = newImage.user.name
-        image.creationData = newImage.creationData
-        image.location = newImage.links.downloadLocation
-        image.loadingCount = Int64(newImage.likes ?? 0)
+        image.image = newImage?.urls?.small
+        image.authorName = newImage?.user?.name
+        image.creationData = newImage?.creationData
+        image.location = newImage?.links?.downloadLocation
+        image.loadingCount = Int64(newImage?.likes ?? 0)
         saveContext()
     }
     
-    func delete(_ image: SaveImage) {
-        viewContext.delete(image)
-        saveContext()
+    func delete(_ image: SaveImage?) {
+        if let image = image {
+            viewContext.delete(image)
+            saveContext()
+        }
     }
 }
